refactor: 게시글 삭제 mutation으로 처리
