+++ src/features/post/ui/PostDeleteButton.tsx
@@ -1,6 +1,6 @@
 import { Trash2 } from "lucide-react"
 
-import { deletePost } from "../../../entities/post/api/deletePost"
+import { useMutationDeletePost } from "../../../entities/post/model/queries/useMutationDeletePost"
 import { Button } from "../../../shared/ui"
 
 interface PostDeleteButtonProps {
@@ -8,17 +8,10 @@ interface PostDeleteButtonProps {
 }
 
 export const PostDeleteButton = ({ postId }: PostDeleteButtonProps) => {
-  const handleDeleteButtonClick = async (postId: number) => {
-    try {
-      const response = deletePost(postId)
-      // setPosts(posts.filter((post) => post.id !== id))
-    } catch (error) {
-      console.error("게시물 삭제 오류:", error)
-    }
-  }
+  const { mutate: deletePostMutation } = useMutationDeletePost()
 
   return (
-    <Button variant="ghost" size="sm" onClick={() => handleDeleteButtonClick(postId)}>
+    <Button variant="ghost" size="sm" onClick={() => deletePostMutation(postId)}>
       <Trash2 className="w-4 h-4" />
     </Button>
   )
